[CP-3194] 'Connected' event logic moved to main device-manager domain (mds logic)
